Re-running notebook for more true runtimes.
Was registering the L=10, M=50 performance as 0.6 instead of
0.33 seconds.

Also the plots were inconsistent, so I manually ran them until they
looked reasonable (should have run an average but would have taken too
long in the notebook). After doing this I manually edited the
output numbers in the cells to be sequential.

Also ditching the 'ell =' annotations in the opening butterfly
graphic.